Lab5/requirements.txt
dlib==19.24.4
fire==0.6.0
fvcore==0.1.5.post20221221
matplotlib==3.9.0
numpy==1.23.5
onnxruntime==1.18.0
opencv_python==4.9.0.80
pandas==2.2.2
Pillow==9.5.0
Requests==2.32.3
scipy==1.13.1
six==1.16.0
smart_path==1.0.1
tabulate==0.9.0
tensorflow==2.16.1
torch==1.11.0
torchgpipe==0.0.7
torchvision==0.12.0
tqdm==4.66.4

